backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.68.1",
    "celery>=5.5.3",
    "fastapi[standard]>=0.117.1",
    "manim>=0.19.0",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
]
